requirements.txt
python-dotenv==1.1.0
google-generativeai==0.8.5
chromadb==1.0.12
streamlit==1.45.1
playwright==1.52.0
